refactor: add GroupTipsCard component
